Introduce prelude in `ab-riscv-interpreter`
