+++ crates/contracts/core/ab-contract-file/src/contract_instruction_prototype.rs
@@ -1,13 +1,4 @@
-use ab_riscv_interpreter::rv64::b::zbb::rv64_zbb_helpers;
-use ab_riscv_interpreter::rv64::b::zbc::rv64_zbc_helpers;
-use ab_riscv_interpreter::rv64::zk::zbkx::rv64_zbkx_helpers;
-use ab_riscv_interpreter::rv64::zk::zkn::zknd::rv64_zknd_helpers;
-use ab_riscv_interpreter::rv64::zk::zkn::zkne::rv64_zkne_helpers;
-use ab_riscv_interpreter::rv64::zk::zkn::zknh::rv64_zknh_helpers;
-use ab_riscv_interpreter::{
-    ExecutableInstruction, ExecutionError, InterpreterState, ProgramCounter,
-    SystemInstructionHandler, VirtualMemory,
-};
+use ab_riscv_interpreter::prelude::*;
 use ab_riscv_macros::{instruction, instruction_execution};
 use ab_riscv_primitives::prelude::*;
 use core::fmt;
