+++ crates/execution/ab-riscv-act4-runner/src/abundance_rv64i_max/interpreter.rs
@@ -1,9 +1,6 @@
 use crate::abundance_rv64i_max::instruction::AbundanceRv64IMaxInstruction;
 use crate::interpreter::{Act4InstructionFetcher, Act4Memory, Act4SystemHandler};
-use ab_riscv_interpreter::v::vector_registers::{
-    VectorRegisterFile, VectorRegisters, VectorRegistersBase, VectorRegistersExt,
-};
-use ab_riscv_interpreter::{CsrError, Csrs, ExecutableInstruction, InterpreterState};
+use ab_riscv_interpreter::prelude::*;
 use ab_riscv_primitives::prelude::*;
 use std::collections::BTreeMap;
