+++ crates/execution/ab-riscv-act4-runner/src/instruction.rs
@@ -1,7 +1,4 @@
-use ab_riscv_interpreter::zicsr::zicsr_helpers;
-use ab_riscv_interpreter::{
-    CsrError, Csrs, ExecutableInstruction, ExecutionError, InterpreterState,
-};
+use ab_riscv_interpreter::prelude::*;
 use ab_riscv_macros::{instruction, instruction_execution};
 use ab_riscv_primitives::prelude::*;
 use std::fmt;
