+++ crates/execution/ab-riscv-act4-runner/src/main.rs
@@ -20,10 +20,7 @@ use crate::abundance_rv32i_max::interpreter::AbundanceRv32IMaxExtState;
 use crate::abundance_rv64i_max::instruction::AbundanceRv64IMaxInstruction;
 use crate::abundance_rv64i_max::interpreter::AbundanceRv64IMaxExtState;
 use crate::interpreter::{Act4InstructionFetcher, Act4SystemHandler};
-use ab_riscv_interpreter::{
-    BasicInt, Csrs, ExecutableInstruction, ExecutionError, FetchInstructionResult,
-    InstructionFetcher, InterpreterState, ProgramCounter, VirtualMemory,
-};
+use ab_riscv_interpreter::prelude::*;
 use ab_riscv_primitives::prelude::*;
 use anyhow::Context;
 use clap::{Parser, ValueEnum};
