+++ crates/execution/ab-riscv-benchmarks/src/host_utils.rs
@@ -5,11 +5,7 @@ use ab_contract_file::ContractInstruction;
 use ab_core_primitives::ed25519::{Ed25519PublicKey, Ed25519Signature};
 use ab_io_type::IoType;
 use ab_io_type::bool::Bool;
-use ab_riscv_interpreter::{
-    BasicInt, ExecutableInstruction, ExecutionError, FetchInstructionResult, InstructionFetcher,
-    InterpreterState, ProgramCounter, ProgramCounterError, SystemInstructionHandler, VirtualMemory,
-    VirtualMemoryError,
-};
+use ab_riscv_interpreter::prelude::*;
 use ab_riscv_primitives::prelude::*;
 use alloc::vec::Vec;
 use core::marker::PhantomData;
